refactor(cli): modularize command dispatch and approval adapter
dispatch.ts becomes a thin entry over dispatch/ (arg tokens, flag
reordering, command tree, usage, unknown command/flag, cli error);
approvalAdapter.ts keeps event orchestration and re-exports while policy,
summaries, event dispatch, and human-input handlers move to approval/.
Deduplicate the bounded-line and line-truncation helpers.
