+++ packages/cli/src/commands/_helpers/dispatch/cliError.ts
@@ -1,0 +1,7 @@
+export function isCliError(error: unknown): error is Error & { code?: string } {
+  return (
+    error instanceof Error &&
+    error.name === 'CLIError' &&
+    typeof (error as { code?: unknown }).code === 'string'
+  );
+}
