+++ packages/cli/src/runtime/approvalAdapter.ts
@@ -1,600 +1,52 @@
-// Third-party imports
-import { structuredPatch } from 'diff';
+// Public entry point for the CLI approval adapter. Internal CLI callers may
+// import the focused modules under ./approval/ directly; this barrel keeps the
+// stable surface for command modules, the TUI, and tests.
 
-// Local imports - runtime
-import { runCoordinatorBridge } from '@agent/runtime/runCoordinators';
 import type { ProgressEventPayloads } from '@eventBus/ProgressEventBus';
-import {
-  agentProposalCategoryLabel,
-  getProposalFileGroups,
-  type AgentProposalPermission,
-  type ApprovalDecision as SharedApprovalDecision,
-} from '@shared/schemas';
 
-// Local imports - tools
-import { handleUserQuestionAction } from '@tools/userQuestion';
-import { handleProgressViewBashApprovalAction } from '@tools/approval/bashApproval';
 import {
   setToolEditApprovalHandler,
   type ToolEditApprovalRequest,
   type ToolEditApprovalResult,
 } from '@tools/approval/toolEditApproval';
-import { handleExternalInquiryAction } from '@tools/inquiry/ExternalInquiryTool';
 
-// Local imports - CLI runtime
+import { isCliDecisionApprovalEvent } from './approvalEvents';
+import { type CliContext } from './cliContext';
+import { writeTextStderr } from './logSinks';
+
 import {
-  isCliDecisionApprovalEvent,
-  type CliDecisionApprovalEvent,
-} from './approvalEvents';
-import { type CliContext, type CliPromptRequest } from './cliContext';
-import { askCliQuestion, writeTextStderr } from './logSinks';
-import { parseUserQuestionAnswer } from './userQuestionAnswer';
+  askApproval,
+  immediateDecision,
+  immediateDecisionForApproval,
+} from './approval/approvalPolicy';
+import {
+  formatRetryRequestMessage,
+  formatToolEditApprovalSummary,
+} from './approval/approvalSummaries';
+import {
+  dispatchApprovalDecision,
+  summarizeApprovalEvent,
+} from './approval/eventDispatch';
+import {
+  handleExternalInquiry,
+  handleUserQuestion,
+} from './approval/humanInputHandlers';
 
-// The headless adapter only ever sets accepted + userMessage, but reusing the
-// host-neutral shape (which also carries optional userQuestionAnswers) keeps a
-// single source of truth for the decision vocabulary across hosts. See
-// docs/proposals/tui-extension-sharing.md (Rung 1).
-export type ApprovalDecision = SharedApprovalDecision;
-
-const TRUNCATED_DIFF_LINE_MARKER = ' … [line truncated]';
-const TOOL_EDIT_APPROVAL_DIFF_MAX_CHARS = 12_000;
-const TOOL_EDIT_APPROVAL_DIFF_MAX_LINES = 80;
-const TOOL_EDIT_APPROVAL_DIFF_MAX_LINE_CHARS = 240;
-const AGENT_PROPOSAL_INSTRUCTION_MAX_CHARS = 6_000;
-const AGENT_PROPOSAL_INSTRUCTION_MAX_LINES = 40;
-const AGENT_PROPOSAL_INSTRUCTION_MAX_LINE_CHARS = 240;
-const AGENT_PROPOSAL_FILE_GROUP_MAX_FILES = 10;
-
-const CLI_PERSONAL_API_RETRY_HINT =
-  'Use `/api personal` in the chat TUI, or press `k` on the retry prompt, to switch to personal API keys.';
-
-const deniedApprovalContexts = new WeakSet<CliContext>();
-const cliPromptQueues = new WeakMap<CliContext, Promise<unknown>>();
-
-export function denyMessage(policy: CliContext['approvalPolicy']): string {
-  return policy === 'ask'
-    ? 'Interactive approval requires a TTY; this CLI run is headless.'
-    : 'Denied by CLI approval policy.';
-}
-
-export function markApprovalDenied(context: CliContext): void {
-  deniedApprovalContexts.add(context);
-}
-
-export function hasCliApprovalDenied(context: CliContext): boolean {
-  return deniedApprovalContexts.has(context);
-}
-
-function isCliMonthlySpendingLimitMessage(
-  message: string | undefined,
-): boolean {
-  return (
-    message?.toLowerCase().includes('monthly spending limit reached') ?? false
-  );
-}
-
-export function isCliApiSwitchableRetry(
-  payload: ProgressEventPayloads['showRetryRequest'],
-): boolean {
-  const details = payload.errorDetails;
-  return (
-    details?.isCredentialExhausted === true &&
-    (details.isRelayError === true ||
-      isCliMonthlySpendingLimitMessage(payload.errorMessage))
-  );
-}
-
-export function appendCliApiSwitchHint(text: string): string {
-  if (!isCliMonthlySpendingLimitMessage(text)) return text;
-  if (text.includes('/api personal')) return text;
-  return [text, CLI_PERSONAL_API_RETRY_HINT].join('\n');
-}
-
-function approvalPromptAllowed(context: CliContext): boolean {
-  return context.approvalPolicy === 'ask' && context.mode === 'interactive';
-}
-
-function enqueueCliPrompt<T>(
-  context: CliContext,
-  prompt: () => Promise<T>,
-): Promise<T> {
-  const previous = cliPromptQueues.get(context) ?? Promise.resolve();
-  const next = previous.catch(() => undefined).then(prompt);
-  cliPromptQueues.set(
-    context,
-    next.catch(() => undefined),
-  );
-  return next;
-}
-
-async function askCliApprovalQuestion(
-  context: CliContext,
-  request: CliPromptRequest,
-): Promise<string> {
-  if (context.approvalPrompt) {
-    return context.approvalPrompt(request);
-  }
-  return askCliQuestion(`${request.summary}\n${request.prompt}`);
-}
-
-export function immediateDecision(
-  context: CliContext,
-): ApprovalDecision | undefined {
-  if (context.approvalPolicy === 'yolo') return { accepted: true };
-  if (approvalPromptAllowed(context)) return undefined;
-  markApprovalDenied(context);
-  return { accepted: false, userMessage: denyMessage(context.approvalPolicy) };
-}
-
-/** Retry requests caused by exhausted credentials or auth failures need user
- *  action (key swap, top-up, re-login). Interactive runs must still surface the
- *  retry panel so the user can switch to personal API keys; non-interactive
- *  and auto-approval modes should not burn the retry budget. */
-function isUnretryableRetryRequest(
-  event: CliDecisionApprovalEvent,
-  payload: ProgressEventPayloads[CliDecisionApprovalEvent],
-): boolean {
-  if (event !== 'showRetryRequest') return false;
-  const details = (payload as ProgressEventPayloads['showRetryRequest'])
-    .errorDetails;
-  if (!details) return false;
-  if (details.isCredentialExhausted) return true;
-  if (details.statusCode === 401 || details.statusCode === 403) return true;
-  return false;
-}
-
-export function immediateDecisionForApproval(
-  event: CliDecisionApprovalEvent,
-  payload: ProgressEventPayloads[CliDecisionApprovalEvent],
-  context: CliContext,
-): ApprovalDecision | undefined {
-  if (isUnretryableRetryRequest(event, payload)) {
-    if (approvalPromptAllowed(context)) return undefined;
-    markApprovalDenied(context);
-    return {
-      accepted: false,
-      userMessage: 'Retry skipped: credential exhausted or unauthorized.',
-    };
-  }
-  return immediateDecision(context);
-}
-
-async function askApproval(
-  context: CliContext,
-  summary: string,
-): Promise<ApprovalDecision> {
-  let answer: string;
-  try {
-    answer = await enqueueCliPrompt(context, () =>
-      askCliApprovalQuestion(context, {
-        kind: 'approval',
-        summary,
-        prompt: 'Approve? [y/N] ',
-      }),
-    );
-  } catch {
-    markApprovalDenied(context);
-    return { accepted: false, userMessage: 'CLI approval prompt failed.' };
-  }
-
-  const normalized = answer.trim().toLowerCase();
-  const accepted = normalized === 'y' || normalized === 'yes';
-  if (!accepted) markApprovalDenied(context);
-  return {
-    accepted,
-    userMessage: accepted ? undefined : 'Rejected from CLI approval prompt.',
-  };
-}
-
-function truncateAgentProposalInstructionLine(line: string): string {
-  if (line.length <= AGENT_PROPOSAL_INSTRUCTION_MAX_LINE_CHARS) return line;
-  const prefixLength = Math.max(
-    0,
-    AGENT_PROPOSAL_INSTRUCTION_MAX_LINE_CHARS -
-      TRUNCATED_DIFF_LINE_MARKER.length,
-  );
-  return `${line.slice(0, prefixLength)}${TRUNCATED_DIFF_LINE_MARKER}`;
-}
-
-function boundedAgentProposalInstructionLines(
-  instruction: string,
-): readonly string[] {
-  const instructionLines = instruction
-    .replaceAll('\r\n', '\n')
-    .replaceAll('\r', '\n')
-    .split('\n');
-  const visibleLines: string[] = [];
-  let visibleChars = 0;
-  let hiddenLineCount = 0;
-
-  for (let index = 0; index < instructionLines.length; index++) {
-    const line = truncateAgentProposalInstructionLine(instructionLines[index]!);
-    const separatorLength = visibleLines.length === 0 ? 0 : 1;
-    const nextVisibleChars = visibleChars + separatorLength + line.length;
-    if (
-      visibleLines.length >= AGENT_PROPOSAL_INSTRUCTION_MAX_LINES ||
-      nextVisibleChars > AGENT_PROPOSAL_INSTRUCTION_MAX_CHARS
-    ) {
-      hiddenLineCount = instructionLines.length - index;
-      break;
-    }
-
-    visibleLines.push(line);
-    visibleChars = nextVisibleChars;
-  }
-
-  if (hiddenLineCount > 0) {
-    visibleLines.push(`… +${hiddenLineCount} instruction lines hidden`);
-  }
-
-  return visibleLines;
-}
-
-function optionalAgentProposalLine(label: string, value: string | undefined) {
-  const trimmed = value?.trim();
-  return trimmed ? [`${label}: ${trimmed}`] : [];
-}
-
-function formatAgentProposalFileGroup(
-  label: string,
-  files: readonly string[],
-): string {
-  const visibleFiles = files.slice(0, AGENT_PROPOSAL_FILE_GROUP_MAX_FILES);
-  const hiddenCount = files.length - visibleFiles.length;
-  const suffix = hiddenCount > 0 ? `, +${hiddenCount} more` : '';
-  return `${label}: ${visibleFiles.join(', ')}${suffix}`;
-}
-
-function agentProposalFileGroupLines(
-  proposal: AgentProposalPermission,
-): string[] {
-  return getProposalFileGroups(proposal).map((group) =>
-    formatAgentProposalFileGroup(group.label, group.files),
-  );
-}
-
-export function formatAgentProposalApprovalSummary(
-  proposal: AgentProposalPermission,
-): string {
-  const instructionLines = boundedAgentProposalInstructionLines(
-    proposal.instruction,
-  );
-  return [
-    `Agent proposal requested: ${proposal.agent} (${agentProposalCategoryLabel(
-      proposal.agentCategory,
-    )})`,
-    `Model: ${proposal.model}`,
-    ...optionalAgentProposalLine(
-      'Working directory',
-      proposal.workingDirectory ?? undefined,
-    ),
-    ...agentProposalFileGroupLines(proposal),
-    'Instruction:',
-    ...instructionLines.map((line) => `  ${line}`),
-  ].join('\n');
-}
-
-function summarizeApprovalEvent<K extends CliDecisionApprovalEvent>(
-  event: K,
-  payload: ProgressEventPayloads[K],
-): string {
-  switch (event) {
-    case 'showBashPermission': {
-      const data = payload as ProgressEventPayloads['showBashPermission'];
-      return `Bash command requested:\n${data.command}`;
-    }
-    case 'showPlanApproval': {
-      const data = payload as ProgressEventPayloads['showPlanApproval'];
-      return `Plan approval requested:\n${JSON.stringify(data.plan, null, 2)}`;
-    }
-    case 'showAgentProposal': {
-      const data = payload as ProgressEventPayloads['showAgentProposal'];
-      return formatAgentProposalApprovalSummary(data);
-    }
-    case 'showRetryRequest': {
-      const data = payload as ProgressEventPayloads['showRetryRequest'];
-      return `Retry requested for ${data.operation}: ${data.errorMessage ?? 'unknown error'}`;
-    }
-    default: {
-      const never: never = event;
-      return String(never);
-    }
-  }
-}
-
-export function formatRetryRequestMessage(
-  payload: ProgressEventPayloads['showRetryRequest'],
-): string {
-  const message = `Retry requested (${payload.operation}): ${payload.errorMessage ?? 'unknown error'}`;
-  if (isCliApiSwitchableRetry(payload)) {
-    return [message, CLI_PERSONAL_API_RETRY_HINT].join('\n');
-  }
-  return message;
-}
-
-function dispatchApprovalDecision<K extends CliDecisionApprovalEvent>(
-  event: K,
-  payload: ProgressEventPayloads[K],
-  decision: ApprovalDecision,
-  options: { writeRejectionToStderr?: boolean } = {},
-): void {
-  const action = decision.accepted ? 'approve' : 'reject';
-  switch (event) {
-    case 'showBashPermission': {
-      const data = payload as ProgressEventPayloads['showBashPermission'];
-      void handleProgressViewBashApprovalAction({
-        requestId: data.requestId,
-        action,
-        feedback: decision.userMessage,
-      });
-      return;
-    }
-    case 'showPlanApproval': {
-      const data = payload as ProgressEventPayloads['showPlanApproval'];
-      runCoordinatorBridge.resolvePlanApproval(data.approvalId, {
-        action,
-        ...(decision.userMessage ? { feedback: decision.userMessage } : {}),
-      });
-      return;
-    }
-    case 'showAgentProposal': {
-      const data = payload as ProgressEventPayloads['showAgentProposal'];
-      runCoordinatorBridge.resolveProposal(data.proposalId, {
-        action,
-        ...(decision.userMessage ? { feedback: decision.userMessage } : {}),
-      });
-      return;
-    }
-    case 'showRetryRequest': {
-      const data = payload as ProgressEventPayloads['showRetryRequest'];
-      if (decision.accepted) {
-        runCoordinatorBridge.triggerRetry(data.streamId);
-        return;
-      }
-      if (options.writeRejectionToStderr) {
-        const summary = summarizeApprovalEvent(event, payload);
-        writeTextStderr(
-          decision.userMessage
-            ? `${summary}\n${decision.userMessage}`
-            : summary,
-        );
-      }
-      runCoordinatorBridge.cancelRetry(data.streamId);
-      return;
-    }
-    default: {
-      const never: never = event;
-      return never;
-    }
-  }
-}
-
-function humanInputDenialFeedback(
-  context: CliContext,
-  yoloMessage: string,
-): string {
-  if (context.approvalPolicy === 'yolo') return yoloMessage;
-  markApprovalDenied(context);
-  return denyMessage(context.approvalPolicy);
-}
-
-function handleExternalInquiry(
-  payload: ProgressEventPayloads['showExternalInquiry'],
-  context: CliContext,
-): void {
-  const threadId = payload.threadId;
-  if (!threadId) {
-    // No persistent thread to address — pre-async legacy payload. Ignore.
-    return;
-  }
-
-  if (!approvalPromptAllowed(context)) {
-    const feedback = humanInputDenialFeedback(
-      context,
-      'External inquiry requires human input; yolo mode cannot synthesize an external answer.',
-    );
-    void handleExternalInquiryAction({ action: 'drop', threadId, feedback });
-    return;
-  }
-
-  void (async () => {
-    let answer: string;
-    try {
-      answer = await enqueueCliPrompt(context, () =>
-        askCliApprovalQuestion(context, {
-          kind: 'externalInquiry',
-          summary: `External inquiry requested:\n${payload.question}`,
-          prompt: 'Answer (blank to skip): ',
-        }),
-      );
-    } catch {
-      markApprovalDenied(context);
-      await handleExternalInquiryAction({
-        action: 'drop',
-        threadId,
-        feedback: 'CLI external inquiry prompt failed.',
-      });
-      return;
-    }
-
-    const trimmed = answer.trim();
-    if (trimmed.length === 0) {
-      await handleExternalInquiryAction({
-        action: 'drop',
-        threadId,
-        feedback: 'External inquiry skipped by user.',
-      });
-      return;
-    }
-    await handleExternalInquiryAction({
-      action: 'submit',
-      threadId,
-      answer: trimmed,
-    });
-  })();
-}
-
-function formatUserQuestionPrompt(
-  payload: ProgressEventPayloads['showUserQuestion'],
-): string {
-  return payload.questions
-    .map((question, index) => {
-      const options = question.options
-        .map((option, optionIndex) => {
-          const description = option.description
-            ? ` - ${option.description}`
-            : '';
-          return `  ${optionIndex + 1}. ${option.label}${description}`;
-        })
-        .join('\n');
-      const multi = question.multiSelect
-        ? ' Select comma-separated numbers.'
-        : '';
-      const free = question.allowFreeText ? ' Or type a custom answer.' : '';
-      return `${index + 1}. ${question.question}\n${options}\n${multi}${free}`;
-    })
-    .join('\n\n');
-}
-
-function handleUserQuestion(
-  payload: ProgressEventPayloads['showUserQuestion'],
-  context: CliContext,
-): void {
-  if (!approvalPromptAllowed(context)) {
-    const feedback = humanInputDenialFeedback(
-      context,
-      'User question requires human input; yolo mode cannot synthesize an answer.',
-    );
-    void handleUserQuestionAction({
-      requestId: payload.requestId,
-      action: 'skip',
-      feedback,
-    });
-    return;
-  }
-
-  void (async () => {
-    const answers: Record<string, string | string[]> = {};
-    try {
-      for (const question of payload.questions) {
-        const answer = await enqueueCliPrompt(context, () =>
-          askCliApprovalQuestion(context, {
-            kind: 'approval',
-            summary: payload.context
-              ? `${payload.context}\n\n${formatUserQuestionPrompt({
-                  ...payload,
-                  questions: [question],
-                })}`
-              : formatUserQuestionPrompt({ ...payload, questions: [question] }),
-            prompt: 'Answer (blank to skip): ',
-          }),
-        );
-        const parsed = parseUserQuestionAnswer(answer, question);
-        if (parsed != null) answers[question.question] = parsed;
-      }
-    } catch {
-      markApprovalDenied(context);
-      await handleUserQuestionAction({
-        requestId: payload.requestId,
-        action: 'skip',
-        feedback: 'CLI user question prompt failed.',
-      });
-      return;
-    }
-
-    const submitted = Object.keys(answers).length > 0;
-    await handleUserQuestionAction({
-      requestId: payload.requestId,
-      action: submitted ? 'submit' : 'skip',
-      answers: submitted ? answers : undefined,
-      feedback: submitted ? undefined : 'User question skipped by user.',
-    });
-  })();
-}
-
-function formatDiffRange(start: number, lineCount: number): string {
-  return lineCount === 1 ? String(start) : `${start},${lineCount}`;
-}
-
-function toolEditDiffLines(
-  request: ToolEditApprovalRequest,
-): readonly string[] {
-  const patch = structuredPatch(
-    request.path,
-    request.path,
-    request.originalContent,
-    request.proposedContent,
-    '',
-    '',
-    { context: 3 },
-  );
-  if (patch.hunks.length === 0) return [];
-
-  return [
-    `--- ${request.path}`,
-    `+++ ${request.path}`,
-    ...patch.hunks.flatMap((hunk) => [
-      `@@ -${formatDiffRange(hunk.oldStart, hunk.oldLines)} +${formatDiffRange(
-        hunk.newStart,
-        hunk.newLines,
-      )} @@`,
-      ...hunk.lines,
-    ]),
-  ];
-}
-
-function truncateDiffLine(line: string): string {
-  if (line.length <= TOOL_EDIT_APPROVAL_DIFF_MAX_LINE_CHARS) return line;
-  const prefixLength = Math.max(
-    0,
-    TOOL_EDIT_APPROVAL_DIFF_MAX_LINE_CHARS - TRUNCATED_DIFF_LINE_MARKER.length,
-  );
-  return `${line.slice(0, prefixLength)}${TRUNCATED_DIFF_LINE_MARKER}`;
-}
-
-function boundedToolEditDiffLines(
-  diffLines: readonly string[],
-): readonly string[] {
-  const visibleLines: string[] = [];
-  let visibleChars = 0;
-  let hiddenLineCount = 0;
-
-  for (let index = 0; index < diffLines.length; index++) {
-    const line = truncateDiffLine(diffLines[index]!);
-    const separatorLength = visibleLines.length === 0 ? 0 : 1;
-    const nextVisibleChars = visibleChars + separatorLength + line.length;
-    if (
-      visibleLines.length >= TOOL_EDIT_APPROVAL_DIFF_MAX_LINES ||
-      nextVisibleChars > TOOL_EDIT_APPROVAL_DIFF_MAX_CHARS
-    ) {
-      hiddenLineCount = diffLines.length - index;
-      break;
-    }
-
-    visibleLines.push(line);
-    visibleChars = nextVisibleChars;
-  }
-
-  if (hiddenLineCount > 0) {
-    visibleLines.push(`… +${hiddenLineCount} diff lines hidden`);
-  }
-
-  return visibleLines;
-}
-
-export function formatToolEditApprovalSummary(
-  request: ToolEditApprovalRequest,
-): string {
-  const header = `Tool edit requested by ${request.sourceTool}: ${request.path}`;
-  const diffLines = toolEditDiffLines(request);
-  if (diffLines.length === 0) {
-    return `${header}\nNo content changes proposed.`;
-  }
-
-  const visibleLines = boundedToolEditDiffLines(diffLines);
-
-  return `${header}\nProposed diff:\n${visibleLines.join('\n')}`;
-}
+export {
+  type ApprovalDecision,
+  appendCliApiSwitchHint,
+  denyMessage,
+  hasCliApprovalDenied,
+  immediateDecision,
+  immediateDecisionForApproval,
+  isCliApiSwitchableRetry,
+  markApprovalDenied,
+} from './approval/approvalPolicy';
+export {
+  formatAgentProposalApprovalSummary,
+  formatRetryRequestMessage,
+  formatToolEditApprovalSummary,
+} from './approval/approvalSummaries';
 
 async function decideToolEdit(
   request: ToolEditApprovalRequest,
